refactored CustomUserDetailsService back to service package, and fixed all imports in DataLoader and WebSecurityConfiguration classes
